refactor: WIP: Update components to use react-router-dom.
