+++ src/Home.css
@@ -1,0 +1,5 @@
+.home {
+  display: grid;
+  grid-template-columns: 1fr;
+  grid-template-rows: 38px 1fr;
+}
