+++ src/components/ErrorBoundary.jsx
@@ -1,0 +1,36 @@
+import { Component } from "react";
+import "./ErrorBoundary.css";
+
+class ErrorBoundary extends Component {
+  constructor() {
+    super();
+
+    this.state = {
+      hasErrored: false,
+    };
+  }
+
+  static getDerivedStateFromError(error) {
+    return { hasErrored: true };
+  }
+
+  componentDidCatch(error, info) {
+    console.log(error);
+  }
+
+  render() {
+    if (this.state.hasErrored) {
+      return (
+        <div className="error-image-overlay">
+          <div className="error-image-container" />
+          <h2 className="error-image-text">Sorry this page is broken</h2>
+        </div>
+      );
+    }
+
+    // eslint-disable-next-line react/prop-types
+    return this.props.children;
+  }
+}
+
+export default ErrorBoundary;
