+++ src/data.json
@@ -455,25 +455,7 @@
       "lunch": "1st"
     }
   },
-  {
-    "id": 47,
-    "contacts": {
-      "id": 47,
-      "lastname": "Jason",
-      "firstname": "Cameron",
-      "phone": "9954",
-      "department": "assist",
-      "email": "cjason@gmail.com",
-      "urlNew": "https://raw.githubusercontent.com/ytrkptl/achsdirectory/main/public/CameronJason.png",
-      "url": "https://dl.airtable.com/.attachments/386eb4b7091a66a2868b448a28088661/0330758f/CameronJason.png",
-      "room": "Security 04",
-      "firstblock": "Planning",
-      "secondblock": "Security 04",
-      "thirdblock": "Security 01",
-      "fourthblock": "Security 04",
-      "lunch": "1st"
-    }
-  },
+
   {
     "id": 40,
     "contacts": {
@@ -510,25 +492,6 @@
       "thirdblock": "E05",
       "fourthblock": "Planning",
       "lunch": "2nd"
-    }
-  },
-  {
-    "id": 44,
-    "contacts": {
-      "id": 44,
-      "lastname": "Justin",
-      "firstname": "Martin",
-      "phone": "9612",
-      "department": "assist",
-      "email": "mjustin@gmail.com",
-      "urlNew": "https://raw.githubusercontent.com/ytrkptl/achsdirectory/main/public/MartinJustin.png",
-      "url": "https://dl.airtable.com/.attachments/6a4339e93c7f50f130cebf2fd1f4f363/baeb1c90/MartinJustin.png",
-      "room": "Security 01",
-      "firstblock": "Security 01",
-      "secondblock": "Security 01",
-      "thirdblock": "Security 04",
-      "fourthblock": "Planning",
-      "lunch": "1st"
     }
   },
   {
@@ -607,25 +570,7 @@
       "lunch": "1st"
     }
   },
-  {
-    "id": 45,
-    "contacts": {
-      "id": 45,
-      "lastname": "Max",
-      "firstname": "Reid",
-      "phone": "9739",
-      "department": "assist",
-      "email": "rmax@gmail.com",
-      "urlNew": "https://raw.githubusercontent.com/ytrkptl/achsdirectory/main/public/ReidMax.png",
-      "url": "https://dl.airtable.com/.attachments/159b5e51bbe8bc0f062e9e7290085cd2/fe888028/ReidMax.png",
-      "room": "Security 02",
-      "firstblock": "Security 02",
-      "secondblock": "Security 02",
-      "thirdblock": "Planning",
-      "fourthblock": "Security 02",
-      "lunch": "2nd"
-    }
-  },
+
   {
     "id": 54,
     "contacts": {
@@ -740,25 +685,7 @@
       "lunch": "2nd"
     }
   },
-  {
-    "id": 46,
-    "contacts": {
-      "id": 46,
-      "lastname": "Piers",
-      "firstname": "Blake",
-      "phone": "9826",
-      "department": "assist",
-      "email": "bpiers@gmail.com",
-      "urlNew": "https://raw.githubusercontent.com/ytrkptl/achsdirectory/main/public/BlakePiers.png",
-      "url": "https://dl.airtable.com/.attachments/87caf9aa9943b292c83a160ab2f8469b/0716d9f2/BlakePiers.png",
-      "room": "Security 03",
-      "firstblock": "Security 03",
-      "secondblock": "Planning",
-      "thirdblock": "Security 03",
-      "fourthblock": "Security 03",
-      "lunch": "3rd"
-    }
-  },
+
   {
     "id": 39,
     "contacts": {
